// src/app/shared/models/about.models.ts
export class About {
  whyChooseTawsileHeader: string;
  whyChooseTawsileDesc: string;

  aboutTawsilaHeader: string;
  aboutTawsilaDesc: string;

  visionHeader: string;
  visionDesc: string;

  missionHeader: string;
  missionDesc: string;

  safetyTitle: string;
  safetyDesc: string;

  rideCheckTitle: string;
  rideCheckDesc: string;

  opinionTitle: string;
  opinionDesc: string;
}